+++ scripts/check-nc-old-new.py
@@ -25,15 +25,16 @@ mounts = ["amlr-gliders-deployments-dev", "swfscesd-glider-deployments-data-out"
 for i in mounts:
     gcp.gcs_mount_bucket(i, str(home / f"mnt-gcs/{i}/"), ro=True)
 
-# deployment_name = "calanus-20241019"; year = "2024"; project="ECOSWIM"
-deployment_name = "amlr08-20220513"; year = "2022"; project="SANDIEGO"
+deployment_name = "calanus-20241019"; year = "2024"; project="ECOSWIM"
+# deployment_name = "amlr08-20220513"; year = "2022"; project="SANDIEGO"
 # deployment_name = "stenella-20250414"
 
 # NOTE: OLD DIR expects old dir structure (.../data/processed-L1/...)
 # Set to True if the old directory structure is the new one (e.g., processed-L0, processed-L1, processed-L3, ancillary-products)
 OLD_USE_NEW = True; OBD_PRE=home / f"tests/{deployment_name}"
-OLD_USE_NEW = False; OBD_PRE=home / "mnt-gcs" / "amlr-gliders-deployments-dev" / project
+# OLD_USE_NEW = False; OBD_PRE=home / "mnt-gcs" / "amlr-gliders-deployments-dev" / project
 OLD_BASE_DIR = OBD_PRE / year / deployment_name
+
 NEW_BASE_DIR = (
     home / "mnt-gcs" / "swfscesd-glider-deployments-data-out"
     # home / f"tests/{deployment_name}"